Move MiniCartBean to shoppingcart package
