+++ shopping-cart/app/shoppingcart/common/CartLineItemBean.java
@@ -3,7 +3,7 @@ package shoppingcart.common;
 import common.contexts.UserContext;
 import common.controllers.ReverseRouter;
 import common.models.ProductAttributeBean;
-import common.models.MiniCartLineItemBean;
+import shoppingcart.MiniCartLineItemBean;
 import common.models.ProductDataConfig;
 import io.sphere.sdk.carts.LineItem;
